move resolvers to editor-manager
