+++ packages/editor/src/browser/editor-frontend-module.ts
@@ -18,10 +18,10 @@ import '../../src/browser/style/index.css';
 import '../../src/browser/language-status/editor-language-status.css';
 
 import { ContainerModule } from '@theia/core/shared/inversify';
-import { CommandContribution, MenuContribution } from '@theia/core/lib/common';
+import { bindContributionProvider, CommandContribution, MenuContribution } from '@theia/core/lib/common';
 import { OpenHandler, WidgetFactory, FrontendApplicationContribution, KeybindingContribution, WidgetStatusBarContribution } from '@theia/core/lib/browser';
 import { VariableContribution } from '@theia/variable-resolver/lib/browser';
-import { EditorManager, EditorAccess, ActiveEditorAccess, CurrentEditorAccess } from './editor-manager';
+import { EditorManager, EditorAccess, ActiveEditorAccess, CurrentEditorAccess, EditorSelectionResolver } from './editor-manager';
 import { EditorContribution } from './editor-contribution';
 import { EditorMenuContribution } from './editor-menu';
 import { EditorCommandContribution } from './editor-command';
@@ -48,6 +48,8 @@ export default new ContainerModule(bind => {
 
     bind(EditorManager).toSelf().inSingletonScope();
     bind(OpenHandler).toService(EditorManager);
+
+    bindContributionProvider(bind, EditorSelectionResolver);
 
     bind(EditorCommandContribution).toSelf().inSingletonScope();
     bind(CommandContribution).toService(EditorCommandContribution);
